Restore `Response#content_type` to the original behavior
Since #35709, `Response#conten_type` returns only MIME type correctly.
It is a documented behavior that this method only returns MIME type, so
this change seems appropriate.
https://github.com/rails/rails/blob/39de7fac0507070e3c5f8b33fbad6fced84d97ed/actionpack/lib/action_dispatch/http/response.rb#L245-L249

But unfortunately, some users expect this method to return all
Content-Type that does not contain charset. This seems to be breaking
changes.

We can change this behavior with the deprecate cycle.
But, in that case, a method needs that include Content-Type with
additional parameters. And that method name is probably the
`content_type` seems to properly.

So I changed the new behavior to more appropriate another
method(`mime_type`) and restored `Response#conten_type` to the original
behavior.

The original `conten_type` method seems to have some problems with
charset parsing (it implicitly expects charset to be at the end). So I
think we should consider the appropriate behavior again later.

Fixes #35709.